moves scoreKeeperSetup directly into NewGame page
