+++ src/styles/NewGame.css
